Ruben feedbacks Part 1
