+++ .gitignore
@@ -22,8 +22,16 @@ pyBeam/lib/_pyBeam.so
 pyBeam/lib/_pyBeamAD.so
 pyBeam/lib/_pyBeamFM.so
 
+pybeam/python/in_out/__init__.pyc
+pybeam/python/in_out/BEAM_config.pyc
+pybeam/python/in_out/ReadInputs.pyc
+pybeam/python/util/__init__.pyc
+pybeam/python/util/switch.pyc
+
+pyBeam/swig/__init__.pyc
 pyBeam/swig/__pycache__/
 pyBeam/swig/pyBeam.py
+pyBeam/swig/pyBeam.pyc
 pyBeam/swig/pyBeamAD.py
 pyBeam/swig/pyBeamAD_wrap.cxx
 pyBeam/swig/pyBeamFM.py
