--- pyBeam/python/in_out/__init__.py~
@@ -1,2 +1,0 @@
-from BEAM_config import BEAMConfig
-from ReadMesh import *
