+++ pyBeam/swig/__init__.py
@@ -1,1 +1,2 @@
 from pyBeam import CBeamSolver, CInput, CProperty
+from Input_parsing import *
