--- tests/OneraM6/BEAM_config.cfg~
@@ -1,80 +1,0 @@
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%                                                                              %
-% BEAM configuration file                                                      %
-% Case description: validation example                                         %
-% Author: _Rocco Bombardieri ________________________________________________  %
-% Institution: _UC3M_________________________________________________________  %
-% Date: _01/02/19_                                                             %
-% File Version                                                %
-%                                                                              %
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-
-% Web & flange thickness [m]
-W_THICKNESS = 0.02
-
-% web height [m]
-W_HEIGHT = 0.40
-
-% flange width [m]
-F_WIDTH = 0.20
-
-% Number of overall nodes along the wing (no collapsed) (n. FEM elements = N_NODES -1 )  %% TO BE REMOVED
-N_NODES = 101  
-
-% Young modulus [GPa]
-Y_MODULUS = 70000000000
-
-% Poisson Ratio
-POISSON = 0.3
-
-%  Beam Density [kg/m^3]
-RHO = 2700
-
-% BEAM Length [m]
-B_LENGTH = 30
-
-% Number of rigid modes to be calculated
-N_DOF = 6  
-
-% Applied load [N]	
-LOAD = 5000
-
-% Follower load flag (0 = NON_FOLLOWER, 1 = FOLLOWER, 2 = APPROX_FOLLOWER) (DEFAULT = 0)
-FOLLOWER_FLAG = 0		
-
-% Number of loadsteps used in the analysis (DEFAULT = 1)
-LOAD_STEPS = 1
-
-% Number of iterations used in the analysis  (DEFAULT = 1)
-N_STRUCT_ITER = 30
-
-% Convergence rate criterium (DEFAULT = 1e-4)
-CONV_CRITERIUM = 0.0001
-
-% MESH file 
-%B_MESH = /media/rocco/290CF0EB732D1122/Adjoint_project/pyBeam/tests/OneraM6/mesh_beam.msh
-B_MESH = ../OneraM6/mesh_beam.msh
-
-% PROPERTY file 
-B_PROPERTY =  /media/rocco/290CF0EB732D1122/Adjoint_project/pyBeam/tests/OneraM6/Property_beam.prt
-
-% ================To be implemented
-
-
-
-% LOAD file (.cfg)
-
-
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-
