--- tests/OneraM6/Property_beam.prt~
@@ -1,3 +1,0 @@
-% Respect the format, properties have to start in line following NPROP
-% Format
-NPROP = 1
