--- tests/OneraM6/mesh_beam.msh~
@@ -1,48 +1,0 @@
-% Respect the format, mesh and conectivity have to start in line following NPOIN and NELEM
-
-NDIM = 3
-
-NPOIN=	20
-0.32236 0.00000 0.00000       
-0.35129 0.06296 0.00000       
-0.38022 0.12593 0.00000       
-0.40914 0.18889 0.00000       
-0.43807 0.25185 0.00000       
-0.46700 0.31482 0.00000       
-0.49593 0.37778 0.00000       
-0.52485 0.44074 0.00000       
-0.55378 0.50371 0.00000       
-0.58271 0.56667 0.00000       
-0.61164 0.62963 0.00000       
-0.64057 0.69259 0.00000       
-0.66949 0.75556 0.00000       
-0.69842 0.81852 0.00000       
-0.72735 0.88148 0.00000       
-0.75628 0.94445 0.00000       
-0.78521 1.00741 0.00000       
-0.81413 1.07037 0.00000       
-0.84306 1.13334 0.00000       
-0.87199 1.19630 0.00000       
-
-% Element type: 1 = line, 5 = triangular
-% ELEM_TYPE  ELEM_NODEID_1  ELEM_NODEID_2  ELEM_PROPERTY
-NELEM=	19
-1  1   2  1
-1  2   3  1
-1  3   4  1
-1  4   5  1
-1  5   6  1
-1  6   7  1
-1  7   8  1
-1  8   9  1
-1  9  10  1
-1  10 11  1
-1  11 12  1
-1  12 13  1
-1  13 14  1
-1  14 15  1
-1  15 16  1
-1  16 17  1
-1  17 18  1
-1  18 19  1
-1  19 20  1
